[TEST][RUST] Preserve the original traversal test layout (#745)
Restore specialized visitor tests and their helpers to their original positions in test_structural_visit.rs, preserving the existing C++ fixture substitutions. Remove the separate fixtures executable and restore registration synchronization before every visitor test.

Validation: all 254 Rust workspace tests pass on Linux; visit and mutate suites pass 100 concurrent rounds. Formatting, ASF header, file type, and whitespace checks pass.
