+++ rust/tvm-ffi/tests/test_structural_visit.rs
@@ -17,18 +17,317 @@
  * under the License.
  */
 
-//! Traversal and reflection tests using types registered by C++ at startup.
-//! Tests that register custom types live in `test_structural_visit_fixtures.rs`,
-//! which Cargo runs in a separate process.
-
 use std::cell::{Cell, RefCell};
+use std::sync::LazyLock;
+
+use tvm_ffi::derive::{Object as DeriveObject, ObjectRef as DeriveObjectRef};
 use tvm_ffi::object::ObjectRef;
-use tvm_ffi::{
-    dispatch, get_type_attr, structural_visit, structural_walk, Any, Array, DLDataType,
-    DLDataTypeCode, DefRegionKind, Error, FieldGetter, Function, Map, Object, ObjectRefCore,
-    Result, String as FfiString, StructuralVisitor, TypeIndex, VisitCallbacks, VisitContext,
-    VisitInterrupt, VisitValue, WalkOrder, WalkResult, RUNTIME_ERROR,
+use tvm_ffi::tvm_ffi_sys::{
+    TVMFFIAny, TVMFFIAnyViewToOwnedAny, TVMFFIByteArray, TVMFFIFieldFlagBitMask, TVMFFIFieldInfo,
+    TVMFFISEqHashKind, TVMFFITypeMetadata, TVMFFITypeRegisterAttr,
 };
+use tvm_ffi::{
+    dispatch, get_type_attr, structural_visit, structural_walk, Any, AnyView, Array, DLDataType,
+    DLDataTypeCode, DefRegionKind, Error, FieldGetter, Function, Map, Object, ObjectArc,
+    ObjectCore, ObjectRefCast, ObjectRefCore, Result, String as FfiString, StructuralVisitor,
+    TypeIndex, VisitCallbacks, VisitContext, VisitInterrupt, VisitValue, WalkOrder, WalkResult,
+    RUNTIME_ERROR,
+};
+
+unsafe extern "C" {
+    fn TVMFFITypeGetOrAllocIndex(
+        type_key: *const TVMFFIByteArray,
+        static_type_index: i32,
+        type_depth: i32,
+        num_child_slots: i32,
+        child_slots_can_overflow: i32,
+        parent_type_index: i32,
+    ) -> i32;
+    fn TVMFFITypeRegisterField(type_index: i32, info: *const TVMFFIFieldInfo) -> i32;
+    fn TVMFFITypeRegisterMetadata(type_index: i32, metadata: *const TVMFFITypeMetadata) -> i32;
+}
+
+#[repr(C)]
+#[derive(DeriveObject)]
+#[type_key = "testing.RustStructuralVisitDefRegion"]
+#[type_final]
+struct RustVisitDefRegionObj {
+    base: Object,
+    recursive: Any,
+    plain: Any,
+    non_recursive: Any,
+    both: Any,
+    ignored: Any,
+}
+
+#[repr(C)]
+#[derive(DeriveObjectRef, Clone)]
+struct RustVisitDefRegion {
+    data: ObjectArc<RustVisitDefRegionObj>,
+}
+
+// The Function-hook tests need to retain the FFI visitor and exercise its
+// lifetime and thread checks. Built-in container hooks do not expose that handle.
+#[repr(C)]
+#[derive(DeriveObject)]
+#[type_key = "testing.RustStructuralVisitHook"]
+#[type_final]
+struct RustVisitHookObj {
+    base: Object,
+    selected: Any,
+    ignored: Any,
+}
+
+#[repr(C)]
+#[derive(DeriveObjectRef, Clone)]
+struct RustVisitHook {
+    data: ObjectArc<RustVisitHookObj>,
+}
+
+#[repr(C)]
+#[derive(DeriveObject)]
+#[type_key = "testing.RustStructuralVisitFailingGetter"]
+#[type_final]
+struct RustVisitFailingGetterObj {
+    base: Object,
+    value: Any,
+}
+
+#[repr(C)]
+#[derive(DeriveObjectRef, Clone)]
+struct RustVisitFailingGetter {
+    data: ObjectArc<RustVisitFailingGetterObj>,
+}
+
+thread_local! {
+    static RETAINED_VISITOR: RefCell<Option<Any>> = const { RefCell::new(None) };
+    static REGISTERED_HOOK_REGIONS: RefCell<Vec<i64>> = const { RefCell::new(Vec::new()) };
+    static PROBE_FOREIGN_THREAD_VISITOR: Cell<bool> = const { Cell::new(false) };
+    static FOREIGN_THREAD_VISITOR_ERROR: RefCell<Option<String>> = const { RefCell::new(None) };
+}
+
+unsafe extern "C" fn clone_any_field(field: *mut std::ffi::c_void, result: *mut TVMFFIAny) -> i32 {
+    TVMFFIAnyViewToOwnedAny(field.cast(), result)
+}
+
+unsafe extern "C" fn clone_any_field_then_fail(
+    field: *mut std::ffi::c_void,
+    result: *mut TVMFFIAny,
+) -> i32 {
+    let code = TVMFFIAnyViewToOwnedAny(field.cast(), result);
+    if code != 0 {
+        return code;
+    }
+    Error::set_raised(&runtime_error(
+        "visit getter failed after writing an owning result",
+    ));
+    -1
+}
+
+fn register_any_field(type_index: i32, name: &'static str, offset: usize, flags: i64) {
+    let field = TVMFFIFieldInfo {
+        name: unsafe { TVMFFIByteArray::from_str(name) },
+        doc: unsafe { TVMFFIByteArray::from_str("Rust structural-visit test field") },
+        metadata: unsafe { TVMFFIByteArray::from_str("") },
+        flags,
+        size: std::mem::size_of::<Any>() as i64,
+        alignment: std::mem::align_of::<Any>() as i64,
+        offset: offset as i64,
+        getter: Some(clone_any_field),
+        setter: std::ptr::null_mut(),
+        default_value_or_factory: TVMFFIAny::new(),
+        field_static_type_index: -1,
+    };
+    assert_eq!(unsafe { TVMFFITypeRegisterField(type_index, &field) }, 0);
+}
+
+fn register_visit_type(type_key: &'static str, total_size: usize, kind: TVMFFISEqHashKind) -> i32 {
+    let type_key = unsafe { TVMFFIByteArray::from_str(type_key) };
+    let type_index = unsafe {
+        TVMFFITypeGetOrAllocIndex(
+            &type_key,
+            -1,
+            Object::TYPE_DEPTH + 1,
+            0,
+            1,
+            Object::type_index(),
+        )
+    };
+    assert!(type_index >= TypeIndex::kTVMFFIDynObjectBegin as i32);
+    let metadata = TVMFFITypeMetadata {
+        doc: unsafe { TVMFFIByteArray::from_str("Rust structural-visit test object") },
+        creator: None,
+        total_size: i32::try_from(total_size).unwrap(),
+        structural_eq_hash_kind: kind as i32,
+    };
+    assert_eq!(
+        unsafe { TVMFFITypeRegisterMetadata(type_index, &metadata) },
+        0
+    );
+    type_index
+}
+
+fn register_function_attr(type_index: i32, name: &'static str, function: Function) {
+    let name = unsafe { TVMFFIByteArray::from_str(name) };
+    let mut value = Any::from(function);
+    assert_eq!(
+        unsafe { TVMFFITypeRegisterAttr(type_index, &name, Any::as_data_ptr(&mut value)) },
+        0
+    );
+}
+
+fn call_visitor_from_foreign_thread(visitor: AnyView<'_>) -> String {
+    // Keep the object alive on this thread. The worker uses a non-owning view,
+    // so it neither transfers nor releases the visitor's reference count.
+    let mut owner = Any::from(visitor);
+    let raw = unsafe { *Any::as_data_ptr(&mut owner) };
+    let type_index = raw.type_index;
+    let object = unsafe { raw.data_union.v_obj } as usize;
+    std::thread::spawn(move || {
+        let mut raw = TVMFFIAny::new();
+        raw.type_index = type_index;
+        raw.data_union.v_obj = object as *mut _;
+        let borrowed = std::mem::ManuallyDrop::new(unsafe { Any::from_raw_ffi_any(raw) });
+        match Function::get_global("ffi.StructuralVisitorVisit")
+            .unwrap()
+            .call_packed(&[AnyView::from(&*borrowed), AnyView::from(&1i64)])
+        {
+            Err(error) => error.message().to_string(),
+            Ok(_) => "foreign-thread visitor call unexpectedly succeeded".to_string(),
+        }
+    })
+    .join()
+    .unwrap()
+}
+
+fn registered_visit_hook(args: &[AnyView<'_>]) -> Result<Any> {
+    assert_eq!(args.len(), 2);
+    if PROBE_FOREIGN_THREAD_VISITOR.with(Cell::get) {
+        let message = call_visitor_from_foreign_thread(args[0]);
+        FOREIGN_THREAD_VISITOR_ERROR.with(|error| error.replace(Some(message)));
+    }
+    RETAINED_VISITOR.with(|retained| {
+        retained.replace(Some(Any::from(args[0])));
+    });
+    let def_region_kind = Function::get_global("ffi.StructuralVisitorDefRegionKind")?
+        .call_packed(&[args[0]])
+        .and_then(i64::try_from)?;
+    REGISTERED_HOOK_REGIONS.with(|regions| regions.borrow_mut().push(def_region_kind));
+
+    let node = RustVisitHook::try_from(args[1])?;
+    Function::get_global("ffi.StructuralVisitorVisit")?
+        .call_packed(&[args[0], AnyView::from(&node.data.selected)])
+}
+
+// The runtime type table leaves registration synchronization to its callers.
+// Every test, including those using only built-in types, must wait for this
+// initializer before reading reflection data: registering a custom __s_visit__
+// hook can relocate the column containing the built-in Array/Map hooks.
+static REGISTER_TEST_TYPES: LazyLock<()> = LazyLock::new(|| {
+    let type_index = register_visit_type(
+        RustVisitHookObj::TYPE_KEY,
+        std::mem::size_of::<RustVisitHookObj>(),
+        TVMFFISEqHashKind::kTVMFFISEqHashKindTreeNode,
+    );
+    register_any_field(
+        type_index,
+        "selected",
+        std::mem::offset_of!(RustVisitHookObj, selected),
+        0,
+    );
+    register_any_field(
+        type_index,
+        "ignored",
+        std::mem::offset_of!(RustVisitHookObj, ignored),
+        0,
+    );
+    register_function_attr(
+        type_index,
+        "__s_visit__",
+        Function::from_packed(registered_visit_hook),
+    );
+
+    let type_index = register_visit_type(
+        RustVisitFailingGetterObj::TYPE_KEY,
+        std::mem::size_of::<RustVisitFailingGetterObj>(),
+        TVMFFISEqHashKind::kTVMFFISEqHashKindTreeNode,
+    );
+    let field = TVMFFIFieldInfo {
+        name: unsafe { TVMFFIByteArray::from_str("value") },
+        doc: unsafe { TVMFFIByteArray::from_str("Fail after producing an owning field value") },
+        metadata: unsafe { TVMFFIByteArray::from_str("") },
+        flags: 0,
+        size: std::mem::size_of::<Any>() as i64,
+        alignment: std::mem::align_of::<Any>() as i64,
+        offset: std::mem::offset_of!(RustVisitFailingGetterObj, value) as i64,
+        getter: Some(clone_any_field_then_fail),
+        setter: std::ptr::null_mut(),
+        default_value_or_factory: TVMFFIAny::new(),
+        field_static_type_index: -1,
+    };
+    assert_eq!(unsafe { TVMFFITypeRegisterField(type_index, &field) }, 0);
+
+    let type_index = register_visit_type(
+        RustVisitDefRegionObj::TYPE_KEY,
+        std::mem::size_of::<RustVisitDefRegionObj>(),
+        TVMFFISEqHashKind::kTVMFFISEqHashKindFreeVar,
+    );
+    for (name, offset, flags) in [
+        (
+            "recursive",
+            std::mem::offset_of!(RustVisitDefRegionObj, recursive),
+            TVMFFIFieldFlagBitMask::kTVMFFIFieldFlagBitMaskSEqHashDefRecursive as i64,
+        ),
+        (
+            "plain",
+            std::mem::offset_of!(RustVisitDefRegionObj, plain),
+            0,
+        ),
+        (
+            "non_recursive",
+            std::mem::offset_of!(RustVisitDefRegionObj, non_recursive),
+            TVMFFIFieldFlagBitMask::kTVMFFIFieldFlagBitMaskSEqHashDefNonRecursive as i64,
+        ),
+        (
+            "both",
+            std::mem::offset_of!(RustVisitDefRegionObj, both),
+            TVMFFIFieldFlagBitMask::kTVMFFIFieldFlagBitMaskSEqHashDefRecursive as i64
+                | TVMFFIFieldFlagBitMask::kTVMFFIFieldFlagBitMaskSEqHashDefNonRecursive as i64,
+        ),
+        (
+            "ignored",
+            std::mem::offset_of!(RustVisitDefRegionObj, ignored),
+            TVMFFIFieldFlagBitMask::kTVMFFIFieldFlagBitMaskSEqHashIgnore as i64,
+        ),
+    ] {
+        register_any_field(type_index, name, offset, flags);
+    }
+});
+
+fn test_prelude() {
+    LazyLock::force(&REGISTER_TEST_TYPES);
+}
+
+fn rust_visit_hook(selected: impl Into<Any>, ignored: impl Into<Any>) -> RustVisitHook {
+    test_prelude();
+    RustVisitHook {
+        data: ObjectArc::new(RustVisitHookObj {
+            base: Object::new(),
+            selected: selected.into(),
+            ignored: ignored.into(),
+        }),
+    }
+}
+
+fn rust_visit_failing_getter(value: impl Into<Any>) -> RustVisitFailingGetter {
+    test_prelude();
+    RustVisitFailingGetter {
+        data: ObjectArc::new(RustVisitFailingGetterObj {
+            base: Object::new(),
+            value: value.into(),
+        }),
+    }
+}
 
 fn runtime_error(message: &str) -> Error {
     Error::new(RUNTIME_ERROR, message, "")
@@ -36,6 +335,7 @@ fn runtime_error(message: &str) -> Error {
 
 #[test]
 fn public_reflection_access_uses_registered_field_and_type_attr() {
+    test_prelude();
     // Keep the existing C++ test library linked so its startup registrations
     // are available even when this test is run by itself.
     assert_eq!(
@@ -72,6 +372,7 @@ fn public_reflection_access_uses_registered_field_and_type_attr() {
 
 #[test]
 fn plain_walk_uses_registered_array_hook() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2, 3]);
     let mut integers = 0;
     assert!(structural_walk(
@@ -90,7 +391,31 @@ fn plain_walk_uses_registered_array_hook() {
 }
 
 #[test]
+fn reflected_getter_releases_partial_result_on_error() {
+    test_prelude();
+    let tracked = FfiString::from("a reference-counted reflected visit field");
+    let root = rust_visit_failing_getter(tracked.clone());
+    let count_before = AnyView::from(&tracked).debug_strong_count();
+
+    let error = match structural_walk(
+        &root,
+        |_value: &VisitValue| WalkResult::Advance,
+        WalkOrder::PreOrder,
+    ) {
+        Ok(_) => panic!("failing getter unexpectedly succeeded"),
+        Err(error) => error,
+    };
+
+    assert_eq!(
+        error.message(),
+        "visit getter failed after writing an owning result"
+    );
+    assert_eq!(AnyView::from(&tracked).debug_strong_count(), count_before);
+}
+
+#[test]
 fn plain_walk_visits_map_values_without_visiting_keys() {
+    test_prelude();
     let root: Map<FfiString, i64> = [(FfiString::from("a"), 1i64), (FfiString::from("b"), 2i64)]
         .into_iter()
         .collect();
@@ -115,7 +440,142 @@ fn plain_walk_visits_map_values_without_visiting_keys() {
 }
 
 #[test]
+fn registered_function_hook_controls_children_interrupts_and_lifetime() {
+    test_prelude();
+    RETAINED_VISITOR.with(|retained| {
+        retained.take();
+    });
+    REGISTERED_HOOK_REGIONS.with(|regions| regions.borrow_mut().clear());
+    let root = rust_visit_hook(11i64, 99i64);
+
+    let mut integers = Vec::new();
+    assert!(structural_walk(
+        &root,
+        |value: &VisitValue| {
+            if let Some(integer) = value.cast::<i64>() {
+                integers.push(integer);
+            }
+            WalkResult::Advance
+        },
+        WalkOrder::PreOrder,
+    )
+    .unwrap()
+    .is_none());
+    // Reflection would visit both fields. The registered hook deliberately
+    // visits only `selected`.
+    assert_eq!(integers, vec![11]);
+    REGISTERED_HOOK_REGIONS.with(|regions| {
+        assert_eq!(regions.borrow().as_slice(), &[DefRegionKind::None as i64]);
+    });
+
+    #[derive(Default)]
+    struct RecordingVisitor {
+        integers: Vec<i64>,
+    }
+    impl StructuralVisitor for RecordingVisitor {
+        fn visit(
+            &mut self,
+            value: &VisitValue,
+            def_region_kind: DefRegionKind,
+        ) -> Result<Option<VisitInterrupt>> {
+            if let Some(integer) = value.cast::<i64>() {
+                self.integers.push(integer);
+            }
+            self.default_visit_children(value, def_region_kind)
+        }
+    }
+    let mut visitor = RecordingVisitor::default();
+    assert!(structural_visit(&root, &mut visitor).unwrap().is_none());
+    assert_eq!(visitor.integers, vec![11]);
+
+    let callback_integers = RefCell::new(Vec::new());
+    assert!(
+        structural_visit(&root, |value: i64, _visitor: &mut VisitContext<'_, ()>| {
+            callback_integers.borrow_mut().push(value)
+        },)
+        .unwrap()
+        .is_none()
+    );
+    assert_eq!(*callback_integers.borrow(), vec![11]);
+
+    test_prelude();
+    let wrapped = RustVisitDefRegion {
+        data: ObjectArc::new(RustVisitDefRegionObj {
+            base: Object::new(),
+            recursive: Any::from(root.clone()),
+            plain: Any::new(),
+            non_recursive: Any::new(),
+            both: Any::new(),
+            ignored: Any::new(),
+        }),
+    };
+    assert!(structural_walk(
+        &wrapped,
+        |_value: &VisitValue| WalkResult::Advance,
+        WalkOrder::PreOrder,
+    )
+    .unwrap()
+    .is_none());
+    REGISTERED_HOOK_REGIONS.with(|regions| {
+        assert_eq!(
+            regions.borrow().last().copied(),
+            Some(DefRegionKind::Recursive as i64)
+        );
+    });
+
+    let outcome = structural_walk(
+        &root,
+        |value: &VisitValue| match value.cast::<i64>() {
+            Some(11) => WalkResult::interrupt_with(FfiString::from("stop")),
+            _ => WalkResult::Advance,
+        },
+        WalkOrder::PreOrder,
+    )
+    .unwrap()
+    .unwrap();
+    assert_eq!(FfiString::try_from(outcome.value).unwrap().as_str(), "stop");
+
+    let retained = RETAINED_VISITOR.with(|retained| retained.take().unwrap());
+    let error = match Function::get_global("ffi.StructuralVisitorVisit")
+        .unwrap()
+        .call_packed(&[AnyView::from(&retained), AnyView::from(&1i64)])
+    {
+        Err(error) => error,
+        Ok(_) => panic!("retained structural visitor unexpectedly remained active"),
+    };
+    assert!(error.message().contains("retained after its active call"));
+}
+
+#[test]
+fn registered_hook_rejects_foreign_thread_visitor_callback() {
+    test_prelude();
+    RETAINED_VISITOR.with(|retained| {
+        retained.take();
+    });
+    FOREIGN_THREAD_VISITOR_ERROR.with(|error| {
+        error.take();
+    });
+
+    let root = rust_visit_hook(11i64, 99i64);
+    PROBE_FOREIGN_THREAD_VISITOR.with(|enabled| enabled.set(true));
+    let result = structural_walk(
+        &root,
+        |_value: &VisitValue| WalkResult::Advance,
+        WalkOrder::PreOrder,
+    );
+    PROBE_FOREIGN_THREAD_VISITOR.with(|enabled| enabled.set(false));
+    assert!(result.unwrap().is_none());
+
+    let message = FOREIGN_THREAD_VISITOR_ERROR.with(|error| error.take().unwrap());
+    assert!(message.contains("invoked from a different thread"));
+    RETAINED_VISITOR.with(|retained| {
+        retained.take();
+    });
+}
+
+#[test]
 fn primitive_values_are_leaves_in_pre_and_post_order() {
+    test_prelude();
     assert!(get_type_attr(TypeIndex::kTVMFFIDataType as i32, "__s_visit__").is_none());
     let dtype = DLDataType::new(DLDataTypeCode::kDLFloat, 32, 1);
 
@@ -174,6 +634,7 @@ fn primitive_values_are_leaves_in_pre_and_post_order() {
 
 #[test]
 fn primitive_fast_path_preserves_none_interrupt_and_error() {
+    test_prelude();
     let mut none_calls = 0;
     assert!(structural_walk(
         &Any::new(),
@@ -209,6 +670,7 @@ fn primitive_fast_path_preserves_none_interrupt_and_error() {
 
 #[test]
 fn registered_map_hook_visits_all_values_without_visiting_keys() {
+    test_prelude();
     // More than 4 entries forces the dense (block + iteration list) layout.
     let root: Map<FfiString, i64> = (0..9)
         .map(|i| (FfiString::from(format!("k{i}")), i as i64))
@@ -235,6 +697,7 @@ fn registered_map_hook_visits_all_values_without_visiting_keys() {
 
 #[test]
 fn interrupt_payload_crosses_map_traversal() {
+    test_prelude();
     let root: Map<FfiString, i64> = [(FfiString::from("a"), 1i64), (FfiString::from("b"), 2i64)]
         .into_iter()
         .collect();
@@ -257,6 +720,7 @@ fn interrupt_payload_crosses_map_traversal() {
 
 #[test]
 fn handler_error_crosses_map_traversal() {
+    test_prelude();
     let root: Map<FfiString, i64> = [(FfiString::from("a"), 1i64)].into_iter().collect();
     let error = match structural_walk(
         &root,
@@ -278,6 +742,7 @@ fn handler_error_crosses_map_traversal() {
 
 #[test]
 fn interrupt_stops_without_running_remaining_callbacks() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2, 3]);
     let mut integers = 0;
     let outcome = structural_walk(
@@ -329,6 +794,7 @@ impl StructuralVisitor for ManualRegionVisitor {
 
 #[test]
 fn manual_child_visit_can_override_def_region() {
+    test_prelude();
     let root = Array::new(vec![7i64, 8]);
     let mut probe = ManualRegionVisitor::default();
     assert!(structural_visit(&root, &mut probe).unwrap().is_none());
@@ -352,6 +818,7 @@ impl GeneratedLeafVisitor {
 
 #[test]
 fn generated_visitor_defaults_unmatched_values() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let mut visitor = GeneratedLeafVisitor::default();
     assert!(structural_visit(&root, &mut visitor).unwrap().is_none());
@@ -393,6 +860,7 @@ impl GeneratedRecursiveVisitor {
 
 #[test]
 fn generated_visitor_can_drive_recursion_through_mut_self() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2, 3]);
     let mut visitor = GeneratedRecursiveVisitor::default();
     let interrupt = structural_visit(&root, &mut visitor).unwrap().unwrap();
@@ -429,6 +897,7 @@ impl GenericDispatchProbe {
 
 #[test]
 fn generated_dispatch_supports_pod_and_ordered_catch_all() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let mut probe = GenericDispatchProbe::default();
     assert!(structural_walk(&root, &mut probe, WalkOrder::PreOrder)
@@ -474,6 +943,7 @@ impl StructuralVisitor for StraddleVisitor {
 
 #[test]
 fn visitor_can_straddle_default_children() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let mut probe = StraddleVisitor::default();
     assert!(structural_visit(&root, &mut probe).unwrap().is_none());
@@ -510,6 +980,7 @@ impl OrderProbe {
 
 #[test]
 fn stateful_structural_walk_supports_post_order() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let mut probe = OrderProbe::default();
     assert!(structural_walk(&root, &mut probe, WalkOrder::PostOrder)
@@ -520,6 +991,7 @@ fn stateful_structural_walk_supports_post_order() {
 
 #[test]
 fn nested_walk_restores_the_outer_active_visitor() {
+    test_prelude();
     let outer = Array::new(vec![10i64, 20]);
     let inner = Array::new(vec![1i64, 2]);
     let mut entered_inner = false;
@@ -557,6 +1029,7 @@ fn nested_walk_restores_the_outer_active_visitor() {
 
 #[test]
 fn interrupt_payload_is_returned_to_the_caller() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let outcome = structural_walk(
         &root,
@@ -577,6 +1050,7 @@ fn interrupt_payload_is_returned_to_the_caller() {
 
 #[test]
 fn handler_errors_include_native_visit_path() {
+    test_prelude();
     let root = Array::new(vec![1i64]);
     let error = match structural_walk(
         &root,
@@ -598,6 +1072,7 @@ fn handler_errors_include_native_visit_path() {
 
 #[test]
 fn visitor_errors_include_native_visit_path() {
+    test_prelude();
     struct FailingVisitor;
 
     impl StructuralVisitor for FailingVisitor {
@@ -636,6 +1111,7 @@ fn visitor_errors_include_native_visit_path() {
 
 #[test]
 fn callback_panics_resume_after_the_registered_hook_returns() {
+    test_prelude();
     let root = Array::new(vec![1i64]);
     let panic = match std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
         structural_walk(
@@ -656,6 +1132,7 @@ fn callback_panics_resume_after_the_registered_hook_returns() {
 
 #[test]
 fn visitor_interrupt_propagates_through_default_children() {
+    test_prelude();
     struct InterruptingVisitor;
 
     impl StructuralVisitor for InterruptingVisitor {
@@ -683,6 +1160,7 @@ fn visitor_interrupt_propagates_through_default_children() {
 
 #[test]
 fn closure_walk_receives_def_region_kind() {
+    test_prelude();
     // C++: StructuralWalk<kPreOrder>(root,
     //          [&](const TVarObj* var, TVMFFIDefRegionKind kind) { ... })
     let root = Array::new(vec![1i64, 2]);
@@ -704,6 +1182,7 @@ fn closure_walk_receives_def_region_kind() {
 
 #[test]
 fn closure_walk_supports_post_order_and_skip() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let mut order_probe = Vec::new();
     assert!(structural_walk(
@@ -743,6 +1222,7 @@ fn closure_walk_supports_post_order_and_skip() {
 // ---------------------------------------------------------------------------
 #[test]
 fn chain_accepts_owned_object_ref_links() {
+    test_prelude();
     let root = Array::new(vec![Array::new(vec![1i64]), Array::new(vec![2i64, 3])]);
     let mut lengths = Vec::new();
     assert!(structural_walk(
@@ -765,6 +1245,7 @@ fn chain_accepts_owned_object_ref_links() {
 
 #[test]
 fn chain_links_may_mix_def_region_arity() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let mut kinds = Vec::new();
     let mut objects = 0;
@@ -791,6 +1272,7 @@ fn chain_links_may_mix_def_region_arity() {
 
 #[test]
 fn chain_links_can_skip_children() {
+    test_prelude();
     let root = Array::new(vec![Array::new(vec![1i64]), Array::new(vec![2i64])]);
     let mut arrays = 0;
     let mut integers = 0;
@@ -816,6 +1298,7 @@ fn chain_links_can_skip_children() {
 
 #[test]
 fn chain_link_errors_include_native_visit_path() {
+    test_prelude();
     let root = Array::new(vec![1i64]);
     let error = match structural_walk(
         &root,
@@ -834,6 +1317,7 @@ fn chain_link_errors_include_native_visit_path() {
 
 #[test]
 fn chain_supports_post_order() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let events = std::cell::RefCell::new(Vec::new());
     assert!(structural_walk(
@@ -870,6 +1354,7 @@ impl ObjectCounter {
 
 #[test]
 fn chain_splices_dispatch_walkers_between_closures() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let mut counter = ObjectCounter::default();
     let mut integers = 0;
@@ -889,6 +1374,7 @@ fn chain_splices_dispatch_walkers_between_closures() {
 
 #[test]
 fn chain_supports_full_arity() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2, 3]);
     let mut integers = Vec::new();
     let mut objects = 0;
@@ -929,6 +1415,7 @@ fn chain_supports_full_arity() {
 
 #[test]
 fn typed_lambda_walks_bare_and_as_single_link_tuple() {
+    test_prelude();
     // A lone typed handler needs no tuple: unmatched values (the array
     // itself) advance normally. The 1-tuple spelling routes through the
     // chain impls instead and must agree.
@@ -960,6 +1447,7 @@ fn typed_lambda_walks_bare_and_as_single_link_tuple() {
 
 #[test]
 fn bare_node_lambda_takes_def_region_kind() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let mut objects = 0;
     assert!(structural_walk(
@@ -1000,6 +1488,7 @@ impl StructuralVisitor for InheritedRegionProbe {
 
 #[test]
 fn def_region_is_inherited_through_containers() {
+    test_prelude();
     let root = Array::new(vec![Array::new(vec![1i64, 2])]);
     let mut probe = InheritedRegionProbe {
         at_root: true,
@@ -1010,7 +1499,116 @@ fn def_region_is_inherited_through_containers() {
 }
 
 #[test]
+fn reflected_field_def_region_reaches_typed_handler() {
+    test_prelude();
+    let root = RustVisitDefRegion {
+        data: ObjectArc::new(RustVisitDefRegionObj {
+            base: Object::new(),
+            recursive: Any::from(1i64),
+            plain: Any::from(2i64),
+            non_recursive: Any::from(3i64),
+            both: Any::from(4i64),
+            ignored: Any::from(5i64),
+        }),
+    };
+    let mut seen = Vec::new();
+    assert!(structural_walk(
+        &root,
+        |_value: i64, kind: DefRegionKind| {
+            seen.push(kind);
+            WalkResult::Advance
+        },
+        WalkOrder::PreOrder,
+    )
+    .unwrap()
+    .is_none());
+    assert_eq!(
+        seen,
+        vec![
+            DefRegionKind::Recursive,
+            DefRegionKind::None,
+            DefRegionKind::NonRecursive,
+            DefRegionKind::NonRecursive,
+        ]
+    );
+}
+
+struct FreeVarClampProbe {
+    at_root: bool,
+    seen: Vec<(&'static str, DefRegionKind)>,
+}
+
+impl StructuralVisitor for FreeVarClampProbe {
+    fn visit(
+        &mut self,
+        value: &VisitValue,
+        def_region_kind: DefRegionKind,
+    ) -> Result<Option<VisitInterrupt>> {
+        if self.at_root {
+            self.at_root = false;
+            let root = value.cast::<Array<ObjectRef>>().unwrap();
+            for index in 0..root.len() {
+                let child = root.get(index).unwrap();
+                if let Some(interrupt) = self.visit_child(&child, DefRegionKind::NonRecursive)? {
+                    return Ok(Some(interrupt));
+                }
+            }
+            return Ok(None);
+        }
+
+        if value.as_node::<RustVisitDefRegionObj>().is_some() {
+            self.seen.push(("free_var", def_region_kind));
+        } else if value.cast::<Array<i64>>().is_some() {
+            self.seen.push(("array", def_region_kind));
+        } else if let Some(integer) = value.cast::<i64>() {
+            self.seen.push((
+                if integer == 6 {
+                    "free_child"
+                } else {
+                    "array_child"
+                },
+                def_region_kind,
+            ));
+        }
+        self.default_visit_children(value, def_region_kind)
+    }
+}
+
+#[test]
+fn non_recursive_region_is_clamped_for_free_var_children_only() {
+    test_prelude();
+    let free_var = RustVisitDefRegion {
+        data: ObjectArc::new(RustVisitDefRegionObj {
+            base: Object::new(),
+            recursive: Any::new(),
+            plain: Any::from(6i64),
+            non_recursive: Any::new(),
+            both: Any::new(),
+            ignored: Any::new(),
+        }),
+    };
+    let free_var: ObjectRef = free_var.try_cast().unwrap();
+    let array: ObjectRef = Array::new(vec![7i64]).try_cast().unwrap();
+    let root = Array::new(vec![free_var, array]);
+    let mut probe = FreeVarClampProbe {
+        at_root: true,
+        seen: Vec::new(),
+    };
+    assert!(structural_visit(&root, &mut probe).unwrap().is_none());
+    assert_eq!(
+        probe.seen,
+        vec![
+            ("free_var", DefRegionKind::NonRecursive),
+            ("free_child", DefRegionKind::None),
+            ("array", DefRegionKind::NonRecursive),
+            ("array_child", DefRegionKind::NonRecursive),
+        ]
+    );
+}
+
+#[test]
 fn nested_tuple_chain_exceeds_flat_arity() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2, 3]);
     let mut integers = Vec::new();
     let mut objects = 0;
@@ -1061,6 +1659,7 @@ fn nested_tuple_chain_exceeds_flat_arity() {
 
 #[test]
 fn nested_tuple_first_match_order_is_flattened() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let mut first = 0;
     let mut second = 0;
@@ -1086,6 +1685,7 @@ fn nested_tuple_first_match_order_is_flattened() {
 
 #[test]
 fn callback_visit_defaults_only_when_no_link_matches() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let integers = Cell::new(0);
     assert!(
@@ -1133,6 +1733,7 @@ fn stateful_visit_integer(value: i64, visitor: &mut VisitContext<'_, StatefulVis
 
 #[test]
 fn stateful_callback_visit_uses_ordinary_mutable_state() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2, 3]);
     let mut visitor = VisitCallbacks::new(
         StatefulVisitStats::default(),
@@ -1157,6 +1758,7 @@ struct StatefulVisitDepth {
 
 #[test]
 fn stateful_callback_visit_reborrows_visitor_during_recursion() {
+    test_prelude();
     let root = Array::new(vec![Array::new(vec![1i64, 2])]);
     let mut visitor = VisitCallbacks::new(
         StatefulVisitDepth::default(),
@@ -1180,6 +1782,7 @@ fn stateful_callback_visit_reborrows_visitor_during_recursion() {
 
 #[test]
 fn callback_visit_can_reenter_the_same_fn_through_visitor() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let visits = Cell::new(0);
     assert!(structural_visit(
@@ -1196,6 +1799,7 @@ fn callback_visit_can_reenter_the_same_fn_through_visitor() {
 
 #[test]
 fn callback_visit_tuple_is_first_match_and_can_interrupt() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2, 3]);
     let fallback = Cell::new(0);
     let interrupted = structural_visit(
@@ -1217,7 +1821,51 @@ fn callback_visit_tuple_is_first_match_and_can_interrupt() {
 }
 
 #[test]
+fn callback_visit_supports_node_links_nested_tuples_and_def_regions() {
+    test_prelude();
+    let root = RustVisitDefRegion {
+        data: ObjectArc::new(RustVisitDefRegionObj {
+            base: Object::new(),
+            recursive: Any::from(1i64),
+            plain: Any::from(2i64),
+            non_recursive: Any::from(3i64),
+            both: Any::from(4i64),
+            ignored: Any::from(5i64),
+        }),
+    };
+    let seen = RefCell::new(Vec::new());
+
+    assert!(structural_visit(
+        &root,
+        (
+            (
+                |_value: f64, _visitor: &mut VisitContext<'_, ()>| {},
+                |_node: &RustVisitDefRegionObj, visitor: &mut VisitContext<'_, ()>| {
+                    assert_eq!(visitor.def_region_kind(), DefRegionKind::None);
+                    visitor.visit_children()
+                },
+            ),
+            |value: i64, visitor: &mut VisitContext<'_, ()>| {
+                seen.borrow_mut().push((value, visitor.def_region_kind()));
+            },
+        ),
+    )
+    .unwrap()
+    .is_none());
+    assert_eq!(
+        *seen.borrow(),
+        vec![
+            (1, DefRegionKind::Recursive),
+            (2, DefRegionKind::None),
+            (3, DefRegionKind::NonRecursive),
+            (4, DefRegionKind::NonRecursive),
+        ]
+    );
+}
+
+#[test]
 fn callback_visit_with_overrides_child_def_region() {
+    test_prelude();
     let root = Array::new(vec![1i64, 2]);
     let seen = RefCell::new(Vec::new());
     assert!(structural_visit(
@@ -1246,6 +1894,7 @@ fn callback_visit_with_overrides_child_def_region() {
 
 #[test]
 fn nested_callback_visit_restores_the_outer_active_visitor() {
+    test_prelude();
     let outer = Array::new(vec![10i64, 20]);
     let inner = Array::new(vec![1i64, 2]);
     let entered_inner = Cell::new(false);
@@ -1274,6 +1923,7 @@ fn nested_callback_visit_restores_the_outer_active_visitor() {
 
 #[test]
 fn callback_visit_panics_resume_and_leave_the_next_run_usable() {
+    test_prelude();
     let root = Array::new(vec![1i64]);
     let entered_callback = Cell::new(false);
     let outcome = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
